Fixed parameter bug and added multi-instance parameters
There was a bug in the parameter system where symbols were required
as the hash indicies, but then strings were being used to check if
the parameters existed in the array. Now uses exclusively symbols
for the parameters.

Also, now instead of only single values, arrays can be passed as
parameters. This is useful with the newly added parameter all-keyword
which can be passed ["hello", "world"] in order to produce the url
?all-keyword=hello&all-keyword=world.